fetch data problem fix
